+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.0.2
+
+Bugfixes:
+
+ - Corregida ruta para los comandos
+ 
+ - Simplificado el chdir inicial
+
+Nuevos comandos:
+
+ - Estado: Da el estado actual de la Raspberry (Versión de firmware, temperatura del procesador y carga del sistema)
+
+ - Ip: Devuelve la ip pública (IPv4) del router
+
+ - Url: Devuelve la url de acceso a la interfaz web de transmission remote
+
 ## 0.0.1
 
 Funciones: